refactor(file-writer): rename writer modules
